requirements.txt
selenium~=4.16.0
python-telegram-bot~=13.0